# pos_billing_openapi.yaml
openapi: 3.0.3
info:
  title: Retail POS Billing & Accounting API
  description: |
    Comprehensive REST API specification for a full-featured Retail POS Billing and Accounting 
    software system. Covers all modules including: POS transactions, inventory, GST/tax compliance,
    accounting, customer management, vendor management, purchase orders, reporting, user management,
    webhooks, and integrations (weighing machines, barcode, WhatsApp/SMS, payment gateways).
    
    Designed for Indian retail requirements including GST, e-Invoice (IRN), e-Way Bill, and UPI.
    Applicable to ERPNext / Frappe custom app implementations.

    ## Authentication
    All endpoints require a Bearer JWT token unless marked as public.
    Obtain a token via `POST /auth/login`.

  version: "1.0.0"
  contact:
    name: Hexonics POS API
    email: dev@hexonics.in
  license:
    name: MIT
servers:
  - url: https://api.yourpos.in/api/v1
    description: Production
  - url: https://staging.yourpos.in/api/v1
    description: Staging
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────
paths:
...
  /payments/bulk-action:
    post:
      tags: [Payments]
      summary: Perform bulk actions on payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, ids]
              properties:
                action: { type: string, enum: [delete, cancel, export] }
                ids: { type: array, items: { type: string } }
      responses:
        '200': { description: Bulk action performed }
...
  /expenses/bulk-action:
    post:
      tags: [Expenses]
      summary: Bulk action on expenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, ids]
              properties:
                action: { type: string, enum: [delete, cancel, export] }
                ids: { type: array, items: { type: string } }
      responses:
        '200':
          description: Bulk action executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  message: { type: string }
...